refactor/all: trivial cleanup
