+++ src/secured_data.rs
@@ -267,7 +267,7 @@ impl SecuredData {
 }
 
 #[cfg(test)]
-mod test {
+mod tests {
     use super::*;
     use tempdir::TempDir;
